Add git_tag_release.sh script
Creating a release consists of a few things. Updating a CHANGELOG.md,
tagging a commit and creating a release branch.

This patch adds a helper script, that wraps around git tag, and updates
the CHANGELOG based on format of keep-a-changelog.

As release branches should really be created by the CI, usually after
successfully running certain tasks in the CI, the script purposely does
not do that.

Also missing are some bad-weather tests, only one happy flow test
scenario is available for now.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>